tests/wpt/tests/html/browsers/the-window-object/open-close/open-features-negative-innerwidth-innerheight.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>HTML: window.open `features`: negative values for legacy `innerwidth`, `innerheight`</title>
<meta name=timeout content=long>
<link rel="help" href="https://html.spec.whatwg.org/multipage/#apis-for-creating-and-navigating-browsing-contexts-by-name">

<!-- user agents are not required to support open features other than `noopener`
     and on some platforms position and size features don't make sense -->
<meta name="flags" content="may" />

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/PrefixedPostMessage.js"></script>
<script>
var windowURL = 'resources/message-opener.html';
var featuresPrefix = `top=0,left=0,`;

// https://html.spec.whatwg.org/multipage/infrastructure.html#rules-for-parsing-integers

setup (() => {
  // Before running tests, open a window using features that mimic
  // what would happen if the features tested here were set to invalid
  // values in later tests.
  // In cases where the value for `innerwidth` or `innerheight` is
  // is less than the browser's minimum allowed value for that dimension,
  // but NOT 0, the value affected will become the browser's minimum allowed value.

  // This should result in a minimally-sized window for later comparison
  var featureString = `${featuresPrefix}innerwidth=1,innerheight=1`;
  var prefixedMessage = new PrefixedMessageTest();
  prefixedMessage.onMessage((data, e) => {
    e.source.close();
    runWindowTests(data);
  });
  var win = window.open(prefixedMessage.url(windowURL), '', featureString);
});

function runWindowTests (baselineDimensions) {

  // Negative values for `innerwidth` should result in a window with minimum
  // valid allowed width
  [ 'innerwidth=-404',
    'innerwidth=-404.5',
    'innerwidth=-404e1'
  ].forEach(feature => {
    async_test(t => {
      var prefixedMessage = new PrefixedMessageTest();
      var featureString = `${featuresPrefix}height=405,${feature}`;
      prefixedMessage.onMessage(t.step_func_done((data, e) => {
        e.source.close();
        assert_equals(data.width, baselineDimensions.width, `"${feature} is negative and should result in a minimally-wide window"`);
      }));
      var win = window.open(prefixedMessage.url(windowURL) + "&expected_innerWidth=" + baselineDimensions.width, '', featureString);
    }, `features "${feature}" should NOT set "width=404"`);
  });

  // Negative values for `innerheight` should result in a window with minimum
  // valid allowed height
  [ 'innerheight=-404',
    'innerheight=-404.5',
    'innerheight=-404e1'
  ].forEach(feature => {
    async_test(t => {
      var prefixedMessage = new PrefixedMessageTest();
      var featureString = `${featuresPrefix}width=404,${feature}`;
      prefixedMessage.onMessage(t.step_func_done((data, e) => {
        e.source.close();
        assert_equals(data.height, baselineDimensions.height, `"${feature} is negative and should result in a minimal-height window"`);
      }));
      var win = window.open(prefixedMessage.url(windowURL) + '&expected_innerHeight=' + baselineDimensions.height, '', featureString);
    }, `features "${feature}" should NOT set "height=404"`);
  });
}

</script>
